feat: 공고 상세 페이지에서 북마크 api 연결 및 컴포넌트 파일 분리
