+++ src/app/(project)/project-recruitment/[id]/page.tsx
@@ -1,23 +1,16 @@
-'use client';
-
-import BottomButton from '@/components/common/bottom-button';
-import Drawer from '@/components/common/drawer';
-import Guide from '@/components/common/guide';
-import Icon from '@/components/common/icon';
-import IconButton from '@/components/common/icon-button';
 import { TagList } from '@/components/common/tag-list';
+import {
+  GuestBottom,
+  HostBottom,
+} from '@/components/project/recruitment/recruitment-bottom';
 import {
   Carousel,
   CarouselContent,
   CarouselDots,
   CarouselItem,
 } from '@/components/ui/carousel';
-import { Textarea } from '@/components/ui/textarea';
-import { GuestProjectGuide } from '@/constants/guide';
-import useDisclosure from '@/hooks/useDisclosure';
 import { getRecruitAnnouncement } from '@/lib/api/project';
 import { cn } from '@/lib/utils';
-import { useRouter } from 'next/navigation';
 import { FC } from 'react';
 
 interface ProjectRecruitmentDetailPageProps {
@@ -42,6 +35,7 @@ const ProjectRecruitmentDetailPage: FC<
     retouchingDescription,
     host,
     conceptPhotoUrls,
+    isBookmarked,
   } = projectData;
 
   return (
@@ -109,107 +103,12 @@ const ProjectRecruitmentDetailPage: FC<
         )}
       >
         {userRole === 'HOST' ? (
-          <HostBottom projectId={projectId} />
+          <HostBottom projectId={projectId} isBookmarked={isBookmarked} />
         ) : (
           <GuestBottom title={title} projectId={projectId} />
         )}
       </div>
     </main>
-  );
-};
-
-const HostBottom = ({ projectId }: { projectId: number }) => {
-  const router = useRouter();
-  return (
-    <>
-      <IconButton
-        icon={<Icon id={'bookmark-icon'} size={24} className="text-gray-40" />}
-      />
-      <IconButton
-        icon={<Icon id={'edit-icon'} size={24} className="text-gray-40" />}
-      />
-      <BottomButton
-        variant={'secondary'}
-        size={'large'}
-        label={'신청자 리스트'}
-        className="w-[222px]"
-        onClick={() => router.push(`/project-management/${projectId}`)}
-        // disabled={}
-      />
-    </>
-  );
-};
-
-const GuestBottom = ({
-  title,
-  projectId,
-}: {
-  title: string;
-  projectId: number;
-}) => {
-  return (
-    <>
-      <IconButton
-        icon={<Icon id={'bookmark-icon'} size={24} className="text-gray-40" />}
-      />
-      <IconButton
-        icon={<Icon id={'share-icon'} size={24} className="text-gray-40" />}
-      />
-      <ApplyDrawer title={title} projectId={projectId} />
-    </>
-  );
-};
-
-const ApplyDrawer = ({
-  title,
-  projectId,
-}: {
-  title: string;
-  projectId: number;
-}) => {
-  const { isOpen, onOpenChange, onClose } = useDisclosure(false);
-
-  return (
-    <Drawer
-      title={'지원하기'}
-      open={isOpen}
-      onOpenChange={onOpenChange}
-      onClose={onClose}
-      trigger={
-        <BottomButton
-          variant={'primary'}
-          size={'large'}
-          label={'프로젝트 신청하기'}
-          className="w-[222px]"
-          // disabled={!isNextEnabled}
-        />
-      }
-      className="pb-0"
-    >
-      <div className={cn('relative flex flex-col gap-3 pb-[74px]')}>
-        <div className="font-body-14 flex items-center gap-2 self-stretch rounded-[8px] border-[1px] border-primary p-2 px-3 text-primary">
-          {title}
-        </div>
-        <div className="flex flex-col gap-2">
-          <h1 className={cn('font-title-16 text-gray-10')}>
-            작가님께 하고싶은 말
-          </h1>
-          <Textarea placeholder="자유롭게 적어주세요." />
-        </div>
-        <Guide title="지원 안내" guides={GuestProjectGuide.general} />
-        <div
-          className={cn(
-            'absolute bottom-0 left-0 flex h-[64px] w-full items-center',
-          )}
-        >
-          <BottomButton
-            variant={'primary'}
-            size={'large'}
-            label={'신청 하기'}
-          />
-        </div>
-      </div>
-    </Drawer>
   );
 };
